sign-xpi: flake8 cleanup
